Do not report transient failures
We don't need to report failures to do with AC or battery level to the LVFS.

+++ libfwupd/fwupd-enums.c
@@ -247,6 +247,8 @@ fwupd_update_state_to_string (FwupdUpdateState update_state)
 		return "success";
 	if (update_state == FWUPD_UPDATE_STATE_FAILED)
 		return "failed";
+	if (update_state == FWUPD_UPDATE_STATE_FAILED_TRANSIENT)
+		return "failed-transient";
 	if (update_state == FWUPD_UPDATE_STATE_NEEDS_REBOOT)
 		return "needs-reboot";
 	return NULL;
@@ -273,6 +275,8 @@ fwupd_update_state_from_string (const gchar *update_state)
 		return FWUPD_UPDATE_STATE_SUCCESS;
 	if (g_strcmp0 (update_state, "failed") == 0)
 		return FWUPD_UPDATE_STATE_FAILED;
+	if (g_strcmp0 (update_state, "failed-transient") == 0)
+		return FWUPD_UPDATE_STATE_FAILED_TRANSIENT;
 	if (g_strcmp0 (update_state, "needs-reboot") == 0)
 		return FWUPD_UPDATE_STATE_NEEDS_REBOOT;
 	return FWUPD_UPDATE_STATE_UNKNOWN;
